requirements.txt

streamlit==1.37.0
ultralytics==8.3.241
opencv-python-headless==4.10.0.84
numpy==1.24.3
Pillow==10.0.1
onnxruntime==1.16.0
plotly==5.18.0